+++ core/apps/congeal_warp.cxx
@@ -241,7 +241,7 @@ main( int argc, char* argv[] )
   functional->InitializeXforms( GridSpacing, GridSpacingExact ); // must do this before downsampling template grid
   const cmtk::Types::Coordinate FinestGridSpacing = GridSpacing / (1<<RefineTransformationGrid);
 
-  const float timeBaselineProcess = cmtk::Timers::GetTimeProcess();
+  const double timeBaselineProcess = cmtk::Timers::GetTimeProcess();
 
   if ( ! DisableOptimization )
     {
@@ -328,7 +328,7 @@ main( int argc, char* argv[] )
     }
 
   // determine and print CPU time (by node, if using MPI)
-  const float timeElapsedProcess = cmtk::Timers::GetTimeProcess() - timeBaselineProcess;
+  const double timeElapsedProcess = cmtk::Timers::GetTimeProcess() - timeBaselineProcess;
 #ifdef CMTK_BUILD_MPI    
   std::vector<float> timeElapsedByNodeProcess( mpiSize );
   MPI::COMM_WORLD.Gather( &timeElapsedProcess, 1, MPI::FLOAT, &timeElapsedByNodeProcess[0], 1, MPI::FLOAT, 0 /*root*/ );
